* Separated create-config from config
